* ADD: Added checking the m_sound._handle() in Release build
* FIX: Using _value_type is not always available, need to call inherited::_value_type.
* EDIT: The w_bool is takes bool value. Down with BOOL! .

+++ code/engine.vc2008/xrGame/random32.h
@@ -8,25 +8,26 @@
 
 #pragma once
 
-class CRandom32 {
+class CRandom32
+{
 private:
-	u32			m_seed;
+	u32 m_seed;
 
 public:
-	IC	u32		seed	()
+IC	u32	seed()
 	{
 		return	(m_seed);
 	}
 
-	IC	void	seed	(u32 seed)
+IC	void seed(u32 seed)
 	{
 		m_seed	= seed;
 	}
 
-	IC	u32		random	(u32 range)
+IC	u32 random(u32 range)
 	{
-		m_seed	= 0x08088405*m_seed + 1;
-		return	(u32(u64(m_seed)*u64(range) >> 32));
+		m_seed = 0x08088405 * m_seed + 1;
+		return	u32(m_seed * range >> 32);
 	}
 };
